refactor(JOE-942): state-io + event-queries; work-store ≤2000
Extract materialization/mutate/sync core (state-io) and event/audit list
APIs (event-queries), then fold remaining private helpers into existing
leaves. Façade work-store.ts ~2954→~1956 LOC; hard budget 3000→2000.
